requirements.txt
torch
torchvision
streamlit
timm
matplotlib
pandas
numpy
tqdm
Pillow